+++ docs/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/),
 
 ### Changed
 
+- Bump frida from 17.16.3 to 17.16.4 by [@dependabot] ([#11])
 - Bump frida from 17.15.5 to 17.16.3 by [@dependabot] ([#10])
 - Bump frida from 17.15.4 to 17.15.5 by [@dependabot] ([#7])
 
@@ -35,6 +36,7 @@ Initial release
 [@repinek]: https://github.com/repinek
 [@dependabot]: https://github.com/dependabot
 
+[#11]: https://github.com/astra1dev/MalumMenu-Android/pull/11
 [#10]: https://github.com/astra1dev/MalumMenu-Android/pull/10
 [#7]: https://github.com/astra1dev/MalumMenu-Android/pull/7
 [#6]: https://github.com/astra1dev/MalumMenu-Android/pull/6